revision fixes, part 1
