+++ EU4toV2/EU4ToV2.vcxproj
@@ -221,6 +221,7 @@
     <ClCompile Include="Source\EU4World\Religions\Religions.cpp" />
     <ClCompile Include="Source\EU4World\World.cpp" />
     <ClCompile Include="Source\FlagUtils.cpp" />
+    <ClCompile Include="Source\Helpers\ReformProperties.cpp" />
     <ClCompile Include="Source\Helpers\TechValues.cpp" />
     <ClCompile Include="Source\main.cpp" />
     <ClCompile Include="Source\Mappers\AdjacencyMapper.cpp" />
@@ -247,6 +248,7 @@
     <ClCompile Include="Source\targa.cpp" />
     <ClCompile Include="Source\V2World\BlockedTechSchools.cpp" />
     <ClCompile Include="Source\V2World\StateMapper.cpp" />
+    <ClCompile Include="Source\V2World\V2UncivReforms.cpp" />
     <ClCompile Include="Source\V2World\Vic2CultureUnion.cpp" />
     <ClCompile Include="Source\V2World\Vic2CultureUnionMapper.cpp" />
     <ClCompile Include="Source\V2World\Vic2Regions.cpp" />
@@ -326,6 +328,7 @@
     <ClInclude Include="Source\EU4World\Religions\Religions.h" />
     <ClInclude Include="Source\EU4World\World.h" />
     <ClInclude Include="Source\FlagUtils.h" />
+    <ClInclude Include="Source\Helpers\ReformProperties.h" />
     <ClInclude Include="Source\Helpers\TechValues.h" />
     <ClInclude Include="Source\Mappers\AdjacencyMapper.h" />
     <ClInclude Include="Source\Mappers\CK2TitleMapper.h" />
@@ -367,6 +370,7 @@
     <ClInclude Include="Source\V2World\V2Relations.h" />
     <ClInclude Include="Source\V2World\V2State.h" />
     <ClInclude Include="Source\V2World\V2TechSchools.h" />
+    <ClInclude Include="Source\V2World\V2UncivReforms.h" />
     <ClInclude Include="Source\V2World\V2World.h" />
     <ClInclude Include="Source\V2World\Vic2CultureUnion.h" />
     <ClInclude Include="Source\V2World\Vic2CultureUnionMapper.h" />
